+++ open_alchemy/helpers/resolve_ref.py
@@ -3,8 +3,8 @@
 import re
 import typing
 
-from openapi_sqlalchemy import exceptions
-from openapi_sqlalchemy import types
+from open_alchemy import exceptions
+from open_alchemy import types
 
 _REF_PATTER = re.compile(r"^#\/components\/schemas\/(\w+)$")