+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # Release Notes
 
-## Version _unreleased_
+## Version 0.8.0 - 2019-11-03
 - Add less verbose initialisation with _init_yaml_ and _init_json_.
-- Remove need for separate models file by exposing _Base_ and constructed models at _openapi_sqlalchemy.models_.
+- Remove need for separate models file by exposing _Base_ and constructed models at _open_alchemy.models_.
+- Update name from OpenAPI-SQLAlchemy to OpenAlchemy
 
 ## Version 0.7.0 - 2019-10-27
 - Add support for Python 3.6.